feat(Chatbot): Add support for high contrast (#711)
Add borders to container that show in high contrast mode. Also fixes an older drawer bug pointed out by @srambach - it was adding a border to the ChatBot on mobile, when drawer styling changes from inline to more of a block layout. We should add it only when we need it on the drawer.